+++ scripts/battlereport/content.js
@@ -6,16 +6,22 @@ let lost = {
     owned: {
         metal: 0,
         mineral: 0,
+        food: 0,
+        energy: 0,
         score: 0,
     },
     allied: {
         metal: 0,
         mineral: 0,
+        food: 0,
+        energy: 0,
         score: 0,
     },
     hostile: {
         metal: 0,
         mineral: 0,
+        food: 0,
+        energy: 0,
         score: 0,
     },
 };
@@ -33,33 +39,42 @@ let brscore = {
 };
 
 document.querySelectorAll(".report .unit").forEach((el) => {
-    const shipName = el.children[0].innerText.trim();
+    const itemId = el.children[0].innerText.trim();
     const owned_before = parseValue(el.children[1].innerText);
     const owned_after = parseValue(el.children[2].innerText);
     const allied_before = parseValue(el.children[3].innerText);
     const allied_after = parseValue(el.children[4].innerText);
     const hostile_before = parseValue(el.children[5].innerText);
     const hostile_after = parseValue(el.children[6].innerText);
-    if (shipName in dgMeta) {
-        brscore.before.owned += owned_before * dgMeta[shipName].score;
-        brscore.after.owned += owned_after * dgMeta[shipName].score;
-        brscore.before.allied += allied_before * dgMeta[shipName].score;
-        brscore.after.allied += allied_after * dgMeta[shipName].score;
-        brscore.before.hostile += hostile_before * dgMeta[shipName].score;
-        brscore.after.hostile += hostile_after * dgMeta[shipName].score;
+    const item = getItemById(itemId);
+    if (item.score) {
+        brscore.before.owned += owned_before * item.score;
+        brscore.after.owned += owned_after * item.score;
+        brscore.before.allied += allied_before * item.score;
+        brscore.after.allied += allied_after * item.score;
+        brscore.before.hostile += hostile_before * item.score;
+        brscore.after.hostile += hostile_after * item.score;
 
         /*
          * I'll use negtive number to emphasis the fact that these are lost resources
          */
-        lost.owned.metal += (owned_after - owned_before) * dgMeta[shipName].metal;
-        lost.owned.mineral += (owned_after - owned_before) * dgMeta[shipName].mineral;
-        lost.owned.score += (owned_after - owned_before) * dgMeta[shipName].score;
-        lost.allied.metal += (allied_after - allied_before) * dgMeta[shipName].metal;
-        lost.allied.mineral += (allied_after - allied_before) * dgMeta[shipName].mineral;
-        lost.allied.score += (allied_after - allied_before) * dgMeta[shipName].score;
-        lost.hostile.metal += (hostile_after - hostile_before) * dgMeta[shipName].metal;
-        lost.hostile.mineral += (hostile_after - hostile_before) * dgMeta[shipName].mineral;
-        lost.hostile.score += (hostile_after - hostile_before) * dgMeta[shipName].score;
+        lost.owned.metal += (owned_after - owned_before) * item.metal;
+        lost.owned.mineral += (owned_after - owned_before) * item.mineral;
+        lost.owned.food += (owned_after - owned_before) * item.food;
+        lost.owned.energy += (owned_after - owned_before) * item.energy;
+        lost.owned.score += (owned_after - owned_before) * item.score;
+
+        lost.allied.metal += (allied_after - allied_before) * item.metal;
+        lost.allied.mineral += (allied_after - allied_before) * item.mineral;
+        lost.allied.food += (allied_after - allied_before) * item.food;
+        lost.allied.energy += (allied_after - allied_before) * item.energy;
+        lost.allied.score += (allied_after - allied_before) * item.score;
+
+        lost.hostile.metal += (hostile_after - hostile_before) * item.metal;
+        lost.hostile.mineral += (hostile_after - hostile_before) * item.mineral;
+        lost.hostile.food += (hostile_after - hostile_before) * item.food;
+        lost.hostile.energy += (hostile_after - hostile_before) * item.energy;
+        lost.hostile.score += (hostile_after - hostile_before) * item.score;
     }
 });
 
@@ -102,25 +117,36 @@ if (container) {
                     </tr>
                     <tr class="unit text-right">
                         <td class="opacBackground text-left">Metal</td>
-                        <td class="opacLightBackground">${formatNumber(lost.owned.metal)}</td>
-                        <td class="opacLightBackground">${formatNumber(lost.allied.metal)}</td>
-                        <td class="opacLightBackground">${formatNumber(lost.hostile.metal)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.owned.metal)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.allied.metal)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.hostile.metal)}</td>
                     </tr>
                     <tr class="unit text-right">
                         <td class="opacBackground text-left">Mineral</td>
-                        <td class="opacLightBackground">${formatNumber(lost.owned.mineral)}</td>
-                        <td class="opacLightBackground">${formatNumber(lost.allied.mineral)}</td>
-                        <td class="opacLightBackground">${formatNumber(lost.hostile.mineral)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.owned.mineral)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.allied.mineral)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.hostile.mineral)}</td>
+                    </tr>
+                     <tr class="unit text-right">
+                        <td class="opacBackground text-left">Food</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.owned.food)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.allied.food)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.hostile.food)}</td>
+                    </tr>
+                    <tr class="unit text-right">
+                        <td class="opacBackground text-left">Energy</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.owned.energy)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.allied.energy)}</td>
+                        <td class="opacLightBackground">${formatNumberInt(lost.hostile.energy)}</td>
                     </tr>
                     <tr class="unit text-right">
                         <td class="opacBackground text-left">Total (Yours + Allied)</td>
                         <td class="opacBackground" colspan="2">
-                            ${formatNumber(
-            lost.owned.metal + lost.owned.mineral + lost.allied.metal + lost.allied.mineral,
-        )}
+                            ${formatNumberInt(lost.owned.metal + lost.owned.mineral + lost.owned.food + lost.owned.energy
+            + lost.allied.metal + lost.allied.mineral + lost.allied.food + lost.allied.energy)}
                         </td>
                         <td class="opacBackground">
-                            ${formatNumber(lost.hostile.metal + lost.hostile.mineral)}
+                            ${formatNumberInt(lost.hostile.metal + lost.hostile.mineral + lost.hostile.food + lost.hostile.energy)}
                         </td>
                     </tr>
                     <tr class="unit text-right">